+++ src/modules/rec-services/watchlater/usage-info.tsx
@@ -18,7 +18,7 @@ export function WatchLaterUsageInfo({ total }: { total: number }) {
   const color: TagColor = 'success'
   const title = `共 ${total} 个视频`
 
-  const { shuffleForWatchLater, addSeparatorForWatchLater, watchlaterNormalOrderSortByAddAtAsc } =
+  const { watchlaterUseShuffle, watchlaterAddSeparator, watchlaterNormalOrderSortByAddAtAsc } =
     useSettingsSnapshot()
   const onRefresh = useOnRefreshContext()
 
@@ -28,11 +28,11 @@ export function WatchLaterUsageInfo({ total }: { total: number }) {
       await delay(100)
       onRefresh?.()
     })()
-  }, [shuffleForWatchLater, addSeparatorForWatchLater, watchlaterNormalOrderSortByAddAtAsc])
+  }, [watchlaterUseShuffle, watchlaterAddSeparator, watchlaterNormalOrderSortByAddAtAsc])
 
   const switchDisplay = (
     <SwitchSettingItem
-      configKey={'shuffleForWatchLater'}
+      configKey={'watchlaterUseShuffle'}
       checkedChildren='随机顺序: 开'
       unCheckedChildren='随机顺序: 关'
       tooltip={<>随机顺序不包括近期添加的视频</>}
@@ -40,7 +40,7 @@ export function WatchLaterUsageInfo({ total }: { total: number }) {
   )
   const checkboxDisplay = (
     <CheckboxSettingItem
-      configKey={'shuffleForWatchLater'}
+      configKey={'watchlaterUseShuffle'}
       label='随机顺序'
       tooltip={
         <>
@@ -52,7 +52,7 @@ export function WatchLaterUsageInfo({ total }: { total: number }) {
   )
   const btnDisplay = (
     <ShuffleSettingsItemFor
-      configKey='shuffleForWatchLater'
+      configKey='watchlaterUseShuffle'
       tooltip='随机顺序不包括近期添加的稍后再看'
     />
   )
@@ -78,7 +78,7 @@ export function WatchLaterUsageInfo({ total }: { total: number }) {
       {/* {checkboxDisplay} */}
       {btnDisplay}
 
-      {!shuffleForWatchLater && (
+      {!watchlaterUseShuffle && (
         <ButtonSettingItem
           configKey='watchlaterNormalOrderSortByAddAtAsc'
           tooltip={